* ./kite:     - Change: In TrackFixedSegment, when the source/target coordinates of the         fixed segment are outside the grid (i.e. the Cell Abutment Box) truncate         them.     - Change: In GCell::addTrackSegment, when looking for overlap between the         to be inserted TrackSegment and Blockage segment, shrink the Blockage         span by one DbU to avoid false overlap when they are just edges connexes.     - Bug: In RoutingEvent::slackenTopology, in the local segment FSM, looping         on the MaximumSlack state. Now also checks the state count.     - Change: In TrackSegment::canMoveUp() & canPivotUp() adds the "reserve"         parameter and uses it in RoutingEvent to prevent over saturated GCells.     - Change: In Configuration, parameters adjustements: expandStep increased to         0.40. Makes RoutingSets biggers but no noticeable slow down.

+++ kite/src/TrackElement.cpp
@@ -2,7 +2,7 @@
 // -*- C++ -*-
 //
 // This file is part of the Coriolis Software.
-// Copyright (c) UPMC/LIP6 2008-2009, All Rights Reserved
+// Copyright (c) UPMC/LIP6 2008-2010, All Rights Reserved
 //
 // ===================================================================
 //
@@ -176,8 +176,8 @@ namespace Kite {
 
 
   bool           TrackElement::canDesalignate             () const { return false; }
-  bool           TrackElement::canPivotUp                 () const { return false; };
-  bool           TrackElement::canMoveUp                  () const { return false; };
+  bool           TrackElement::canPivotUp                 ( float ) const { return false; };
+  bool           TrackElement::canMoveUp                  ( float ) const { return false; };
   bool           TrackElement::canDogLeg                  () { return false; };
   bool           TrackElement::canDogLeg                  ( Interval ) { return false; };
   bool           TrackElement::canDogLegAt                ( GCell*, bool allowReuse ) { return false; };
